third_party/WebKit/LayoutTests/editing/execCommand/insertNewLineInQuotedContent-outside-quote.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_selection(
    '^foo<svg></svg>|',
    selection => {
      const document = selection.document;
      document.designMode = 'on';
      document.execCommand('InsertNewlineInQuotedContent');
    },
    '^foo|<svg></svg>');
}, 'InsertNewlineInQuotedContent outside of quote element should not do nothing');
</script>
